requirements.txt
certifi==2024.2.2
influxdb-client==1.41.0
numpy~=1.26.4
pandas~=2.2.1
python-dateutil==2.9.0.post0
python-dotenv~=1.0.1
pytz==2024.1
reactivex==4.0.4
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
tqdm~=4.66.2
matplotlib~=3.8.3
seaborn~=0.13.2
typer~=0.9.0
joblib~=1.4.0
scipy~=1.13.0
torch~=2.3.0
scikit-learn~=1.4.2
lightning~=2.2.4
torchmetrics~=1.4.0
rootutils~=1.0.7
omegaconf~=2.3.0